Move authErrorHandler.ts to middlewares folder
