+++ src/middlewares/authErrorHandler.ts
